Wiiu Port
FakeEmitter/FakeJit/VertexDecoderFake buildfix.

big-endian fixes.

add wiiu build files.

add wiiu build files.

initial wiiu port.

fix IR interpreter on big-endian machines.

fix VertexDecoder/SoftGpu on big-endian.

Revert "Goodbye PowerPC, you can always be resurrected when the time comes"

This reverts commit a54e0cf2445a461df173137ca26df84ab6c80d98.

WiiU: improve exception handler.

PpcJit now works on WiiU, sort of ...

PpcJit: disable instructions not available on WiiU.

add GPU_GX2, doesn't work yet ...

WiiU: add a basic GX2 validation layer.

WiiU: some small improvements to GX2_GPU. still not functional.

WiiU: hardcode some shaders for now.

big endian fix in Loaders.cpp

WiiU: more improvements to GPU_GX2. cube.elf now renders correctly.

GPUCommon: fix freeze on big-endian.

Wii-U: start generating shaders dynamically.

Wii-U: add an ubershader. a lot of scenes render correctly now.

Wii-U: add support for Hardware T&L.

Wii-U: add missing compile flags. fixes savestates.

Wii-U: copy tiled surface data correctly. fixes screenshots.

Wii-U: use a push buffer for uniforms.

README: include temporary manual ffmpeg submodule updating (#1)

Manually update ffmpeg submodule for now, see (https://github.com/hrydgard/ppsspp/pull/11001#issuecomment-386886497)

more big-endian fixes.

Wii-U: misc.

Wii-U: add stencil upload shader.

Wii-U: add missing support for analog sticks input.

Wii-U: implement CopyFramebufferImage.

Wii-U: GX2: fix color attributes.

Swap.h: get rid of most overloads. they weren't necessary and were
having some really bad side-effects.(like preventing type promotions,
eg: result of `swapped<short> * int` was being clamped to short)

fix SceAtracIdInfo struct on big-endian.

simplify Swap.h

Wii-U: register axis input only on change.

keep decoded/mixed audio as little endian in PSP memory, to allow
emulated code to process it correctly.

more big-endian fixes.

Wii-U: update shaders.

another round of big-endian fixes.

Wii-U: misc.

libkirk: replace some magic offsets with struct member access. this will
also allow using the scalar_storage_order attribute to enforce endianess
(gcc only).

a bit more big-endian fun.

Wii-U: simulate a square range for analog input.

Wii-U: build fix.

more big-endian fixes.

Wii-U: fix webserver.

build fix for devkitPPC r38.

(WIIU)set toolchain file automatically. use std=c++17.

build fix after rebase.

debug build fix.

texture_atlas big endian fixes.

WiiU: GX2 fixes after rebase.

WiiU: exception handler tweaks.

nits.

big endian fixes.

derp.

WiiU: thin3d: use correct index type.

misc.

update the ffmpeg module

WiiU: fix flipped display

PrxDecrypter: big endian fixes.

change ext/wiiu into a submodule.

big endian ...

WiiU: GX2:
-add COLOR_2D_RB_SWIZZLE shader.
-use correct pitch and endianess with TextureDesc.initDataCallback;

Wiiu: GX2: use correct component swizzle.

WiiU: GX2: need to use the backbuffer ptr here.

WiiU: GX2: no need to endian swap here, fixes screenshots.

WiiU: refactor *ShaderGeneratorGX2.

WiiU: plug leak.

WiiU: rename shader_info.h
update wiiu submodule.

WiiU: add a basic shader disassembler.

WiiU: improve shader generators. use fallback ubershader for unsupported
shaderIDs.

Mpeg: avoid flagging the mpeg data as invalid when MpegDemux::demux is
called with a small read size.

WiiU: GX2 fix scissor (was being limited to display res instead of
current framebuffer res).

WiiU: refactors.

WiiU: GX2Framebuffer: fall back to MEM2 when running out of MEM1 memory.

WiiU: refactors.

WiiU: update *ShaderGenerator

big endian.

printf big endian fixes.

misc.

misc.

wiiu: update wiiu.cmake

Swap.h: add parameter pack templates to automatically cast swap_t<>
values to their basic types when passed as parameters to *_LOG calls.
plus some refactors.

use little endian types for keys in PrxDecrypter.cpp

WiiU: misc.

Revert "printf big endian fixes."

This reverts commit 2334beba1e186f95eecc97edbb2e5483999700e7.

+++ Core/Dialog/SavedataParam.cpp
@@ -649,7 +649,7 @@ int SavedataParam::LoadSaveData(SceUtilitySavedataParam *param, const std::strin
 	std::string filename = GetFileName(param);
 	std::string filePath = dirPath + "/" + filename;
 	s64 readSize;
-	INFO_LOG(SCEUTILITY,"Loading file with size %u in %s",param->dataBufSize,filePath.c_str());
+	INFO_LOG(SCEUTILITY,"Loading file with size %u in %s",(u32)param->dataBufSize,filePath.c_str());
 	u8 *saveData = nullptr;
 	int saveSize = -1;
 	if (!ReadPSPFile(filePath, &saveData, saveSize, &readSize)) {
